Move StoreConfig to bin crate
